moved a set of functions from DccdConvertExportApp to dccd in dccdConvertPackage, added a few unit tests for the functions in dccd
