+++ src/main/assembly/bin.xml
@@ -57,7 +57,7 @@
         </fileSet>
         <fileSet>
             <directory>${project.basedir}/src/main/assembly/dist</directory>
-            <outputDirectory />
+            <outputDirectory/>
             <excludes>
                 <exclude>bin</exclude>
             </excludes>
