+++ src/test/resources/debug-config/application.properties
@@ -1,1 +1,2 @@
-projects =./data/projects
+projects=./data/projects
+data=./data
