+++ src/test/resources/debug-config/logback-service.xml
@@ -12,8 +12,8 @@
         </encoder>
     </appender>
     <root level="warn">
-        <appender-ref ref="FILE" />
-        <appender-ref ref="STDOUT" />
+        <appender-ref ref="FILE"/>
+        <appender-ref ref="STDOUT"/>
     </root>
-    <logger name="nl.knaw.dans.easy" level="trace" />
+    <logger name="nl.knaw.dans.easy" level="trace"/>
 </configuration>
