+++ src/test/resources/debug-config/logback.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <configuration>
-	<appender name="FILE"
-			  class="ch.qos.logback.core.FileAppender">
-		<file>data/dccd-convert-export-cmd.log</file>
-		<encoder>
-			<pattern>[%thread] %-5level %logger{15} - %msg%n</pattern>
-		</encoder>
-	</appender>
-	<root level="warn">
-		<appender-ref ref="FILE" />
-	</root>
-	<logger name="nl.knaw.dans.easy" level="trace" />
+    <appender name="FILE"
+              class="ch.qos.logback.core.FileAppender">
+        <file>data/dccd-convert-export-cmd.log</file>
+        <encoder>
+            <pattern>[%thread] %-5level %logger{15} - %msg%n</pattern>
+        </encoder>
+    </appender>
+    <root level="warn">
+        <appender-ref ref="FILE"/>
+    </root>
+    <logger name="nl.knaw.dans.easy" level="trace"/>
 </configuration>
