+++ src/test/scala/nl.knaw.dans.dccd.convert/CustomMatchers.scala
@@ -17,20 +17,21 @@ package nl.knaw.dans.dccd.convert
 
 import java.io.File
 
+import org.scalatest.matchers.{ MatchResult, Matcher }
+
 import scala.io.Source
 
-import org.scalatest.matchers.{MatchResult, Matcher}
-
 /** Does not dump the full file but just the searched content if it is not found.
-  *
-  * See also <a href="http://www.scalatest.org/user_guide/using_matchers#usingCustomMatchers">CustomMatchers</a> */
+ *
+ * See also <a href="http://www.scalatest.org/user_guide/using_matchers#usingCustomMatchers">CustomMatchers</a> */
 trait CustomMatchers {
   class ContentMatcher(content: String) extends Matcher[File] {
     def apply(left: File): MatchResult = {
       def trimLines(s: String): String = s.split("\n").map(_.trim).mkString("\n")
+
       MatchResult(
         trimLines(Source.fromFile(left).mkString).contains(trimLines(content)),
-        s"$left did not contain: $content" ,
+        s"$left did not contain: $content",
         s"$left contains $content"
       )
     }
