+++ src/test/scala/nl.knaw.dans.dccd.convert/ReadmeSpec.scala
@@ -36,9 +36,9 @@ class ReadmeSpec extends FlatSpec with Matchers with CustomMatchers {
   }
 
   "options in help info" should "be part of README.md" in {
-    val lineSeparators = s"(${System.lineSeparator()})+"
-    val options = helpInfo.split(s"${lineSeparators}Options:$lineSeparators")(1)
-    options.trim.length shouldNot be (0)
+    val lineSeparators = s"(${ System.lineSeparator() })+"
+    val options = helpInfo.split(s"${ lineSeparators }Options:$lineSeparators")(1)
+    options.trim.length shouldNot be(0)
     new File("README.md") should containTrimmed(options)
   }
